Wine Quality Prediction
In this notebook, some features that affect the quality of a wine are given.

The following steps are applied in the project in order:

The dataset is imported.

It is observed whether there is missing data in the dataset and then necessary arrangements have been made.

Statistical (descriptive) analysis and then visualization of the data were made for better understanding before modeling.

Outliers of the data have been removed.

First, standardization processes were applied to the data and then is divided into train and test.

Machine learning models were created and model success was determined.